requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
transformers
onnxruntime
scipy
wandb
Pillow
